[NEAR-44] Refactor: 도메인별로 테스트 디렉토리 구조 구성
- Move auth related tests to `tests/auth/`
- Move user related tests to `tests/users/`
- Move core/security tests to `tests/core/`
- Move chat tests to `tests/chat/`
- Add `__init__.py` to new test packages for better discovery
