+++ tests/core/test_security.py
